+++ src/main/java/com/aivle/bookapp/dto/UserResponseDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class UserResponseDto {
-    private Long usersId;
+    private Integer usersId;
     private String loginId;
     private String name;
     private Integer gubun;
